GuiServer: Add tables

+++ Ease/src/Ease.hpp
@@ -36,7 +36,7 @@
 template <typename T>
 using Reference = std::shared_ptr<T>;
 
-typedef size_t ResourceID;
+typedef uint64_t ResourceID;
 typedef uint32_t EntityID;
 
 namespace Ease {
